EBrun is working - until the next bug

+++ code/zerocache.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import os
 import numpy as np
-
+from script_utils import show_output
 
 def extract_zero_df(stack_df, zero_string="0|0|0|0|0"):
     """
@@ -40,10 +40,14 @@ def load_zero_df(zero_path):
     load_files = [os.path.join(zero_path, f"zero.{i}.csv") for i in load_file_counter]
     zero_dfs = []
     for f in load_files:
-        if os.path.isfile(f):
+        try:
             zdf = pd.read_csv(f, sep="\t")
-            # show_output(f"Loading zero cache file {os.path.basename(f)}", multi=True)
             zero_dfs.append(zdf)
+            # show_output(f"Loading zero cache file {os.path.basename(f)}", multi=True)
+        except:
+            show_output(f"{f} could not be loaded", color="warning", time=False)
+            # show_output(f"Loading zero cache file {os.path.basename(f)}", multi=True)
+            
     # if nothing could be found
     if len(zero_dfs):
         zero_df = pd.concat(zero_dfs).drop_duplicates("D")
@@ -101,15 +105,23 @@ def collapse_zeros(zero_path, zero_condense_factor=13):
     reduces all zero_files to zero.0.csv
     """
     # get all the zero files
+    show_output(f"Collapsing all zero files in {zero_path} into zero.0.csv")
     zero_files = [
         os.path.join(zero_path, file)
         for file in os.listdir(zero_path)
-        if os.path.isfile(os.path.join(zero_path, file))
+        if os.path.isfile(os.path.join(zero_path, file)) and "zero" in file
     ]
     zdfs = []
+    if len(zero_files) == 0:
+        show_output(f"No zerofiles found in {zero_path}!")
+        return
     for file in zero_files:
-        print(file)
-        zdfs.append(pd.read_csv(file, sep="\t"))
+        show_output(f"Cleaning up zero file {file}", time=False)
+        try:
+            zdf = pd.read_csv(file, sep="\t")
+            zdfs.append(zdf)
+        except:
+            show_output(f"{file} could not be loaded", color="warning", time=False)
         os.remove(file)
     # concat and drop duplicates
     zero_df = pd.concat(zdfs).drop_duplicates("D")
@@ -121,4 +133,5 @@ def collapse_zeros(zero_path, zero_condense_factor=13):
     zero_df = zero_df.drop_duplicates("D").sort_values("D")
 
     zero_df.to_csv(os.path.join(zero_path, "zero.0.csv"), sep="\t", index=False)
+    show_output(f"Written collapsed zerofile ({len(zero_df.index)} lines) to {zero_path}/zero.0.csv")
     return zero_df
